back/requirements.txt
argon2-cffi==20.1.0
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==4.1.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.7
click==8.0.1
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.0
debugpy==1.3.0
decorator==5.1.0
defusedxml==0.7.1
Django==3.2.8
django-allauth==0.46.0
django-appconf==1.0.5
django-cors-headers==3.10.0
django-extensions==3.1.3
django-imagekit==4.1.0
django-seed==0.3.1
django-summernote==0.8.20.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
entrypoints==0.3
Faker==8.10.3
Flask==2.0.1
idna==3.3
inflection==0.5.1
ipykernel==6.0.2
ipython==7.27.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.2
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
oauthlib==3.1.1
packaging==21.0
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
pilkit==2.0
Pillow==8.4.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
psycopg2==2.9.1
pycparser==2.20
Pygments==2.10.0
PyJWT==1.7.1
pyparsing==3.0.3
pyrsistent==0.18.0
python-dateutil==2.8.2
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.3
pywin32==301
pywinpty==1.1.3
PyYAML==5.4.1
pyzmq==22.1.0
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
Send2Trash==1.7.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
terminado==0.10.1
testpath==0.5.0
text-unidecode==1.3
toposort==1.7
tornado==6.1
traitlets==5.1.0
uritemplate==4.1.1
urllib3==1.26.7
urlopen==1.0.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
